requirements/prod.txt
django
djangorestframework
markdown
django-filter
django-oauth-toolkit
